clean up code

+++ src/main/java/com/seat/esg/service/MessageService.java
@@ -1,7 +1,9 @@
 package com.seat.esg.service;
 
 import com.seat.esg.domain.Message;
+import com.seat.esg.domain.SeatStatus;
 import com.seat.esg.repository.MessageRepository;
+import com.seat.esg.repository.SeatRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,17 +16,17 @@ import java.util.List;
 public class MessageService {
 
     private final MessageRepository messageRepository;
+    private final SeatRepository seatRepository;
 
     @Transactional
-    public Long createAwayRequestMessage(int seatNum) {
+    public void createAwayRequestMessage(int seatNum) {
         Message message = Message.createMessage(seatNum);
         messageRepository.save(message);
-        return message.getId();
     }
 
     @Transactional
     public void deleteClearSeatMessage(int seatNum) {
-        if (!messageRepository.findBySeatNum(seatNum).isEmpty()) {
+        if (!messageRepository.findBySeatNum(seatNum).isEmpty() && !(seatRepository.findOneBySeatNum(seatNum).getStatus() == SeatStatus.AWAY)) {
             messageRepository.removeBySeatNum(seatNum);
         }
     }
